// frontend/src/lib/generated/models/ImagesResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * seaport.com
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Image } from './Image';
import {
    ImageFromJSON,
    ImageFromJSONTyped,
    ImageToJSON,
    ImageToJSONTyped,
} from './Image';
import type { ResponseMetadata } from './ResponseMetadata';
import {
    ResponseMetadataFromJSON,
    ResponseMetadataFromJSONTyped,
    ResponseMetadataToJSON,
    ResponseMetadataToJSONTyped,
} from './ResponseMetadata';

/**
 * 
 * @export
 * @interface ImagesResponse
 */
export interface ImagesResponse {
    /**
     * 
     * @type {Array<Image>}
     * @memberof ImagesResponse
     */
    data: Array<Image>;
    /**
     * 
     * @type {ResponseMetadata}
     * @memberof ImagesResponse
     */
    metadata: ResponseMetadata;
}

/**
 * Check if a given object implements the ImagesResponse interface.
 */
export function instanceOfImagesResponse(value: object): value is ImagesResponse {
    if (!('data' in value) || value['data'] === undefined) return false;
    if (!('metadata' in value) || value['metadata'] === undefined) return false;
    return true;
}

export function ImagesResponseFromJSON(json: any): ImagesResponse {
    return ImagesResponseFromJSONTyped(json, false);
}

export function ImagesResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ImagesResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'data': ((json['data'] as Array<any>).map(ImageFromJSON)),
        'metadata': ResponseMetadataFromJSON(json['metadata']),
    };
}

export function ImagesResponseToJSON(json: any): ImagesResponse {
    return ImagesResponseToJSONTyped(json, false);
}

export function ImagesResponseToJSONTyped(value?: ImagesResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'data': ((value['data'] as Array<any>).map(ImageToJSON)),
        'metadata': ResponseMetadataToJSON(value['metadata']),
    };
}