// frontend/src/lib/generated/models/ImageLog.ts
/* tslint:disable */
/* eslint-disable */
/**
 * seaport.com
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ImageLog
 */
export interface ImageLog {
    /**
     * 
     * @type {string}
     * @memberof ImageLog
     */
    data: string;
}

/**
 * Check if a given object implements the ImageLog interface.
 */
export function instanceOfImageLog(value: object): value is ImageLog {
    if (!('data' in value) || value['data'] === undefined) return false;
    return true;
}

export function ImageLogFromJSON(json: any): ImageLog {
    return ImageLogFromJSONTyped(json, false);
}

export function ImageLogFromJSONTyped(json: any, ignoreDiscriminator: boolean): ImageLog {
    if (json == null) {
        return json;
    }
    return {
        
        'data': json['data'],
    };
}

export function ImageLogToJSON(json: any): ImageLog {
    return ImageLogToJSONTyped(json, false);
}

export function ImageLogToJSONTyped(value?: ImageLog | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'data': value['data'],
    };
}